Rename c2ast files to C~ convention; remove unused basicTest.cpp
- cfileinfo.{h,cpp} → CFileInfo.{h,cpp}
- ctoken.{h,cpp} → CToken.{h,cpp}
- clexer.{h,cpp} → CLexer.{h,cpp}
- cpreprocessor.{h,cpp} → CPreprocessor.{h,cpp}
- cparser.{h,cpp} → CParser.{h,cpp}
- Remove stale src/c2ast/basicTest.cpp (tests live in test/c2ast-tester/)
- Update all #include statements and CMakeLists.txt accordingly
