--- doc/memo.md
@@ -1,30 +1,0 @@
-# モジュールの方針
-	javascript方式取り込みたいものとファイルを指定。
-
-# インポートの方針
-	宣言用のソースはなく、元ソースのastのキャッシュから宣言のみ取り込む。
-	ast再生時に、宣言部分が生成される。
-	インポート文が来たら、すでにast生成されてないか確認して
-	されてなかったら、astを、生成し、キャッシュする。。
-	循環をしないため、インポートはスタックされ、で親がきたらそれ以上はネストしてインポートしない。
-	それで問題が、出たら、エラーとする。
-
-# キャッシュフォルダ
-	astのキャッシュは、テンポラリーに作成される。
-	グローバルキャッシュと、プロジェクトキャッシュがある.
-	ソースの親ディレクトリを探索し、.palanがあった場合、そこのast用のキャッシュを
-	使い、なければグローバル(ユーザーの下)を使う。
-
-# 循環依存関係の解決
-	基本はC++と同じで、名前のみ宣言、定義無し構文で
-	回避する。ポインタのみ。
-	インポート前に、お互い、名前のみ宣言する。
-
-# グローバルモジュール
-
-# Cインクルードファイル
-  C言語のファイルはインポートではなく、インクルードとして扱う。
-  つまり、インクルードの宣言が現れた場所に、指定されたファイルのastがインクルードされる。
-  インクルード時に指定があれば、その名前空間に所属し、
-  使用の際は、その名前空間を指定する必要がある。
-
