+++ src/c2ast/CPreprocessor.cpp
@@ -8,10 +8,10 @@
 #include <boost/assert.hpp>
 
 using namespace std;
-#include "cfileinfo.h"
-#include "ctoken.h"
-#include "clexer.h"
-#include "cpreprocessor.h"
+#include "CFileInfo.h"
+#include "CToken.h"
+#include "CLexer.h"
+#include "CPreprocessor.h"
 
 // utilities
 class IfInfo {
